fix: Set headers at runtime (same approach)
